social-network-hw3/user_service/requirements.txt
fastapi
uvicorn
pydantic
email-validator>=2.0.0
SQLAlchemy
psycopg2-binary
python-jose
passlib[bcrypt]
httpx==0.23.0
bcrypt<4.0.0
